pyproject.toml
[tool.poetry]
name = "jarvis-backend"
version = "0.1.0"
description = "JARVIS — Voice-first AI agent backend"
authors = ["Rohit Kale"]
readme = "README.md"
packages = [
    { include = "shared" },
    { include = "api" },
    { include = "agent" },
    { include = "mcp_gateway" },
    { include = "cli_bridge" },
    { include = "rag" },
]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
"uvicorn[standard]" = "^0.32.0"
livekit-agents = "^1.0.0"
livekit-plugins-deepgram = "^1.0.0"
livekit-plugins-silero = "^1.0.0"
livekit-plugins-cartesia = "^1.0.0"
livekit-plugins-anthropic = "^1.0.0"
anthropic = "^0.42.0"
supabase = "^2.11.0"
"redis[hiredis]" = "^5.2.0"
pydantic-settings = "^2.7.0"
httpx = "^0.28.0"
"python-jose[cryptography]" = "^3.3.0"
python-multipart = "^0.0.18"
mcp = "^1.3.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-asyncio = "^0.24.0"
httpx = "^0.28.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
